pyproject.toml
[tool.poetry]
name = "brighter"
version = "0.1.0"
description = "A simple library for data-wrangling with GCP support"
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
apache-airflow = "^2.6.2"
apache-airflow-providers-google = "^10.3.0"
pandas = "^2.0.3"
pandas-gbq = "^0.19.2"
apache-airflow-providers-http = "^4.4.2"
apache-airflow-providers-slack = {extras = ["http"], version = "^7.3.1"}


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
